Allow signing X.509 certificates with Ed25519 keys
With OpenSSL, `MessageDigest::null()` can be used when signing with
Ed25519 keys, but that digest type isn't supported by BoringSSL.
Instead, a null pointer needs to be passed to the signing functions.

Add variants of the signing functions to the X.509 builders that don't
take a digest and pass a null pointer to boringssl.

Test: build, TH, WIP code
Change-Id: I4e0cee237b7608dc134aa29e68c58f9f25a93bf6