Import upstream fixes to type-safety flaws in rust-openssl
This imports:
https://github.com/sfackler/rust-openssl/commit/ed6e1ae859f40c64fb9eb0b01f9e92f8ac6ce14c
https://github.com/sfackler/rust-openssl/commit/6fb228bebb8b09942242c2741e255c408cdbd0ef

These patches are needed to build with the latest BoringSSL.
rust-openssl uses an unsound bindings strategy that is not type-safe.
While upstream rust-openssl have been fixing a few of the surface-level
issues (though not the underlying design flaw), Android's copy is
behind, so we need to manually import patches to unblock updates.

Test: treehugger
Change-Id: Ic8915ff05ea5de62c25a866c1c2dbb01ccab9015