Reduces visibility of rust-openssl to only its direct dependents
This CL restricts the visibility of rust-openssl, so that it is
only available to the packages that currently depend on it.
This is so that no new dependencies on rust-openssl can be introduced.

This is part of an effort to remove all dependencies on rust-openssl,
and migrate the existing ones to use bssl-crypto instead. Once a package's
dependence on rust-openssl has been migrated, it will be removed from
the list of packages to which rust-openssl must be visible.

Test: build
Change-Id: Ifc526f0bed5ffc54a7c59e9962f0204420e11366